Исправлены замечания. Тип данных в базе изменен на NUMERIC. а в коде на BigDecimal

+++ src/main/java/ru/alexandrov/currencychecker/dao/model/BoxPlotData.java
@@ -3,13 +3,15 @@ package ru.alexandrov.currencychecker.dao.model;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class BoxPlotData {
-    private float median25;
-    private float median50;
-    private float median75;
-    private float min;
-    private float max;
+    private BigDecimal median25;
+    private BigDecimal median50;
+    private BigDecimal median75;
+    private BigDecimal min;
+    private BigDecimal max;
     private String message;
 }
